Renamed some types and cleaned up some clippy lints.
Changed BlockSpi to AcquiredSdCard.
Changed SdMmcSpi to SdCard.

+++ src/fat/volume.rs
@@ -149,7 +149,7 @@ impl FatVolume {
             }
             FatSpecificInfo::Fat32(_fat32_info) => {
                 // FAT32 => 4 bytes per entry
-                let fat_offset = cluster.0 as u32 * 4;
+                let fat_offset = cluster.0 * 4;
                 this_fat_block_num = self.lba_start + self.fat_start.offset_bytes(fat_offset);
                 let this_fat_ent_offset = (fat_offset % Block::LEN_U32) as usize;
                 volume_mgr
